+++ common/src/main/java/dev/ftb/packcompanion/api/client/pause/AdditionalPauseProvider.java
@@ -1,12 +1,7 @@
 package dev.ftb.packcompanion.api.client.pause;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import net.minecraft.client.gui.components.Widget;
-import net.minecraft.client.gui.components.events.GuiEventListener;
-import net.minecraft.client.gui.narration.NarratableEntry;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.List;
 
 /**
  * @implNote Please do NOT try and modify the screen. This api is provided to avoid nasty hacks